GitHub issue template: Link to troubleshooting at the top
The main purpose of looking at the troubleshooting page is to help the
user to solve their issue themselves. If they are able to do that, they
don't need to fill out the rest of the form, so lets point to the
troubleshooting doc at the top.

The troubleshooting doc also explains how to enable debug logging. We
could also ask our users to do that and try to reproduce the issue with
debug logging, but that might be asking a bit much for non-technical
users, and is often not necessary to solve the issue, so lets instead
ask for that in a comment when we ascertain that we do need debug logs.