nss/integration-tests: Use lower-case names for pre-checking the user
Now everything in authd expects the users to be lower case, and we
should not rely on the broker policies.

As per this, just always pass to the broker the user name in lower case
when doing the pre-check so that there's no room such issues.

Without this, an user that tries to log in for the first time using a
capitalized name in SSH would not be allowed, unless the broker does
this kind of check.