userutils: Run the whole test inside bubblewrap
Move the helpers tools to run tests in bubblewrap to a more central
place (as they will surely be useful in future for other tests too).

Then run the whole userutils tests inside bubble wrap.

All this is needed because in future we may want to be able to run
tests in an environment where most of the /etc directory has
read-only access, while we are still allowed to mock some of the
files.

And centralizing this allows us to just do such adjustments inside the
test that is launched in the bubblewrap environment, rather than on its
caller, simplifying the logic overall.

As per this we have also few other extra advantages:
 - Be able to easily add and run multiple tests inside the bwrap environment
 - Have a generic way to test inside a such environment (that will be useful
   in future for other integration tests)
 - Run all these tests in parallel