userutils: Use libc standard locking mechanism for shadow password
When modifying any of the passwd file we are supposed to take a lock,
while using per-database .lock files works well enough for testing is
still prone to races, and systemd just ignores it because of this
reason.

So let's just use the standard way and call lckpwdf()/ulckpwdf() key
calls pairs.

Add more tests to cover the case too and an override mechanism to be
used only for testing.